Net_URL2 2.2.0 has switched to classmap autoloader
the require_once statement of it has been been made conditional because
the file is not part of the include-path any longer as it has been removed.

The class_exists check triggers the autoloader. In case there is no
autoloader PEAR include path configuration is expected and require_once
works as known with the standard confgured PEAR include path.